Moved data_node module to conversion directory.
